### 2023-03-20 #5273 @editorialbot
[PRE REVIEW]: msmhelper: A Python package for Markov state modeling of protein dynamics
**Submitting author:** @braniii (<a href="http://orcid.org/0000-0002-2863-2646">Daniel Nagel</a>)
**Repository:** https://github.com/moldyn/msmhelper
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.2
