### 2023-03-20 #5273 @braniii
Based on their reviewed papers I would suggest the following reviewers:
- amritagos
- luthaf
- hmacdope
- anotherjoshsmith
- rasbt

And these base the joss reviewer list:
- yuxuanzhuang
- taoliu032
- lorenzo-rovigatti
- dotsdl

Hope it helps.
