### 2023-03-27 #5273 @braniii
Dear @lucydot,

I wanted to thank you for volunteering to be the editor of my article. If there is anything I can do to assist you during this process, please don't hesitate to reach out. I am more than happy to answer any questions you may have or provide additional information to help facilitate the pre-review and the review process.

Best regards,
Daniel Nagel
