### 2023-03-28 #5273 @lucydot
Hello @dotsdl, @taoliu032, @yuxuanzhuang - are you available to review this submission for JOSS?

I can see that for a couple of you it would be your first review. I am always very happy to answer any questions about the review process you may have, as it is quite different than other journals - it is often more of an on-going discussion between author, reviewers, editor and our bot (!), which we can all learn from.
