### 2023-04-13 moldyn/msmhelper#26 @yuxuanzhuang
JOSS Review--Documentation
This is part of the review process for https://github.com/openjournals/joss-reviews/issues/5339

There's a few links broken in the github page

- [https://github.com/moldyn/msmhelper#usage](https://github.com/moldyn/msmhelper#usage)
    - [x] linking issue with API
    - [x] reference issue with **following section**
- [https://github.com/moldyn/msmhelper#hummer-szabo-projection](https://github.com/moldyn/msmhelper#hummer-szabo-projection)
    - [x] linking issue with **tutorial**
- [https://moldyn.github.io/msmhelper/benchmark/](https://moldyn.github.io/msmhelper/benchmark/)
    - [x] Download doesn’t work

**Updated May 3**
- [https://moldyn.github.io/msmhelper/tutorials/msm/](https://moldyn.github.io/msmhelper/tutorials/msm/)
    - [x] where λ i is the i -th eigenvalue.
        - of what?
- [https://moldyn.github.io/msmhelper/tutorials/timescales/](https://moldyn.github.io/msmhelper/tutorials/timescales/)
    - [x] raw latex in the webpage
        - e.g. t^\text{wt}
    - [x] unfinished sentence
        - Here we see, that choosing a long lag time
- [https://moldyn.github.io/msmhelper/tutorials/hummerszabo/](https://moldyn.github.io/msmhelper/tutorials/hummerszabo/)
    - [x] raw latex in the webpage
    - [x] typo: neglegtable, evry
