### 2023-04-13 moldyn/msmhelper#27 @yuxuanzhuang
JOSS Review--Performance (WIP)
This is part of the review process for https://github.com/openjournals/joss-reviews/issues/5339

Thanks for bringing this wonderful package to the MSM world! I am wondering if you have considered conducting a benchmark comparison between msmhelper and Deeptime (https://deeptime-ml.github.io/latest/index.html). It would be fascinating to see how they compare. However, I understand if this is not a critical aspect of this package review.
