### 2023-04-13 moldyn/msmhelper#27 @braniii
That is an interesting question. I've added a few benchmarks and surprisingly, deeptime is sometimes significantly faster than Pyemma and almost as fast as msmhelper.

But I guess, one should keep in mind that while msmhelper focuses only on estimating msms, deeptime is a much more versatile package focusing on machine learning approaches.
