+++ mkdocs.yml
@@ -106,6 +106,7 @@ plugins:
 - mkdocs-jupyter:
     include: ["*.ipynb"]
     execute: true
+    include_source: true
     allow_errors: false
     ignore_h1_titles: true
 - mkdocstrings:
