### 2023-04-17 cef4100
Prevent parsing newlines in python as breaks.
