+++ mkdocs.yml
@@ -66,7 +66,6 @@ markdown_extensions:
   - footnotes
   - mkdocs-click
   - md_in_html
-  - nl2br
   - toc:
       permalink: true
 
