### 2023-04-17 0eb747f
Fix bug of returning microstate index trajectories instead of macrostate trajectories, #30.
