+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+
+
+## [1.0.3] - 2023-04-17
 ### Bugfix 🐛:
 - Fix that `mh.LumpedStateTraj.index_trajs` and `mh.LumpedStateTraj.index_trajs_flatten` return now index trajectories corresponding to the macrostate trajectories
 
@@ -162,7 +165,8 @@ Chapman-Kolmogorov test
 - Initial release
 
 
-[Unreleased]: https://github.com/moldyn/msmhelper/compare/v1.0.2...main
+[Unreleased]: https://github.com/moldyn/msmhelper/compare/v1.0.3...main
+[1.0.3]: https://github.com/moldyn/msmhelper/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/moldyn/msmhelper/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/moldyn/msmhelper/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/moldyn/msmhelper/compare/v0.6.2...v1.0.0
