+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,34 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior, including a minimal working example.
+```
+import msmhelper as mh
+...
+# prepare data
+...
+# run msmhelper
+mh.function(...)
+# get error
+```
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Runtime information:**
+ - OS: [e.g. Linux]
+ - Version [e.g. v1.0.3]
+ - Source [PyPI, conda-forge, or source]
+
+**Additional context**
+Add any other context about the problem here.
