### 2023-04-24 moldyn/msmhelper#32 @lorenzo-rovigatti
JOSS review: about the paper
This is part of the review process for https://github.com/openjournals/joss-reviews/issues/5339

Hi and thanks for the work you have put into this nice package. Here is a list of comments related to the paper:

- [x] line 62: "Monte chain" -> "Markov chain" (I guess?)
- [ ] panels 1a, 1c and 1e look *extremely* similar to panels and figures of the Nagel et al (2023) paper. While this is not an issue per se, it may become a problem depending on the journal the foregoing paper will be published into. Please make sure that you'll have the rights to re-use the figure (or change the figure so that it won't clash with the journal license).
- [x] About the figure: while I appreciate brevity, I think here additional details about the quantities shown in the figures would help an interested (but maybe not too knowledgeable, or just inexperienced) reader
