### 2023-04-25 moldyn/msmhelper#32 @braniii
Thank you for taking the time to review our paper and for providing helpful feedback. We have addressed the first issue raised in your comments and corrected "Monte chain" to "Markov chain" in line 62. We agree that the third point is a good suggestion and will work to provide additional details about the quantities shown in the figures to improve the clarity of the paper. We appreciate the reviewer for raising the issue of potential copyright infringement related to the similarity of the figures in panels 1a, 1c, and 1e with those in Nagel et al (2023). We acknowledge that the origin of the similarity is due to the use of msmhelper, which is a crucial tool for generating the figures in our paper. We will take steps to address this concern.
