### 2023-05-03 moldyn/msmhelper#26 @yuxuanzhuang
Hi! I've added with a few more comments on the documentations
