### 2023-05-03 moldyn/msmhelper#33 @yuxuanzhuang
JOSS Review--Code
This is part of the review process for https://github.com/openjournals/joss-reviews/issues/5339

- state indexing
    - [x] I am a bit confused about how the state is indexed/reindexed. e.g. in [estimate_markov_model](https://github.com/moldyn/msmhelper/blob/5380ef761cb63f1d8e1bb4f28fd1e1087dd95314/src/msmhelper/msm/msm.py#L28). It is stated that "The states should start from zero" but I guess that's not necessary?

- ~~[plot_ck_test](https://github.com/moldyn/msmhelper/blob/5380ef761cb63f1d8e1bb4f28fd1e1087dd95314/src/msmhelper/plot/_ck_test.py#L21)~~
 ~~unfinished in the documentation~~
