### 2023-05-15 6ceea18
Fix issues with waiting time tutorial, #26
