### 2023-05-15 5297884
Fix typos and rephrase section about alternative methods, #26
