### 2023-05-15 moldyn/msmhelper#26 @braniii
Thank you for bringing this issue to my attention. For now, I will mark it as resolved. However, if you continue to encounter the same LaTeX error, please don't hesitate to reopen this issue and provide more details about the error and the browser you are using. Thank you again for your feedback and helping me improve the documentation!
