### 2023-05-15 86b52e2
Clearify and improve the documentation of estimate_markov_model, #33
