### 2023-05-15 moldyn/msmhelper#33 @braniii
Thank you again for bringing these inconsistencies in the documentation to my attention. I apologize for any confusion that may have caused. You are correct that the states can be any integer and it is also possible to skip states. The error you noticed was due to a copy-paste error from an internal function that was optimized for performance. I appreciate your attention to detail and your feedback, which helps me improve the quality of this project.
