### 2023-05-15 9db98cb
Improve the docstring of the CKtest, #33
