### 2023-05-15 9d5da12
Include plot submodule to the module overview.
