### 2023-05-15 moldyn/msmhelper#33 @braniii
Could you please provide more information about what you mean by "unfinished" in the second point of your previous feedback? In any case, I have updated the docstring of the chapman_kolmogorov_test function to make it clearer.
