### 2023-05-16 0807b13
Fix typo in CK test.
