### 2023-05-16 moldyn/msmhelper#33 @braniii
I understand your query now. The notation [path.to.object][] is utilized in mkdocstrings syntax for the automatic generation of markdown references in the format [path.to.object](correct.html.path). This is a nice feature of the mkdocstrings package, which greatly aids in the creation of automated mkdocs documentations.

I've noted that you've taken off the 'Work in Progress' label. Consequently, I'll consider this issue resolved for the moment. However, if you encounter any additional issues, please don't hesitate to reopen this one. Your continued input is always appreciated.
