### 2023-05-17 moldyn/msmhelper#36 @lorenzo-rovigatti
JOSS review: examples (tutorials) and testing
This is part of the review process for https://github.com/openjournals/joss-reviews/issues/5339

Hi again! My last comments are about examples and testing:

- [x] The *tutorials* section of the documentation is very well done and easy to follow. My only suggestion is to write somewhere (in the README and perhaps also in the tutorials home page of the docs) that the jupyter notebooks are available in the `docs/tutorials` folder
- [x] Very minor suggestion: It may be convenient to some users to have a local copy of the documentation, so why not adding a way of generating the docs html from the downloaded repo, which should amount to installing the requirements and running `python3 -m mkdocs build` or something similar.
- ~[ ] If I install all the testing requirements and run pytest, I get the following error at the end: `FAIL Required test coverage of 90% not reached. Total coverage: 0.00%`. Can you look into it?~
