### 2023-05-17 f2183a3
Add short discussion of the figure, #32
