+++ paper.bib
@@ -191,3 +191,15 @@ eprint = {https://doi.org/10.1021/ct200463m}
   eprint = {https://wires.onlinelibrary.wiley.com/doi/pdf/10.1002/wcms.1343},
   year = {2018}
 }
+@article{Diez22,
+  author = {Diez, Georg and Nagel, Daniel and Stock, Gerhard},
+  title = {Correlation-Based Feature Selection to Identify Functional Dynamics in Proteins},
+  journal = {J. Chem. Theory Comput.},
+  volume = {18},
+  number = {8},
+  pages = {5079-5088},
+  year = {2022},
+  doi = {10.1021/acs.jctc.2c00337},
+  note ={PMID: 35793551},
+  URL = {https://doi.org/10.1021/acs.jctc.2c00337}
+}
