### 2023-05-17 moldyn/msmhelper#32 @braniii
Concerning the third point, the model in Nagel et al. (2023) has been updated for the final publication in JCTC. Hence, there should be no need for changing the figures here. We have also spotted another typo and added a short discussion of the figure.
