### 2023-05-17 3a75c2a
Fix python version to 3.10 for now due to release of 3.11, #32
