### 2023-05-17 1827efa
Add instruction on downloading tutorials and on how to build for offline use, #36
