### 2023-05-17 cc55c2b
Fix invisible symbole.
