### 2023-05-17 moldyn/msmhelper#36 @braniii
Regarding the first two points. We have added the the FAQ a description on how to build the docs locally. Further, we have added the section [documentation and tutorials](https://github.com/moldyn/msmhelper#documentation-and-tutorials) to the README. There we describe what the documentation contains, how the notebooks can be downloaded and how to compile the docs for local use.
