### 2023-05-17 moldyn/msmhelper#36 @lorenzo-rovigatti
About the testing bit: I didn't follow the docs thoroughly, so that for some reason `pytest` did run but did not yield the expected result. Sorry about that, I'll edit the first post!
