### 2023-05-17 #5339 @lorenzo-rovigatti
Hi @lucydot , my review is done! I have open a couple of issues and raised some questions which have been addressed by @braniii . As far as I'm concerned, msmhelper and the accompanying paper are ready to be published in JOSS!
