### 2023-05-18 #5339 @taoliu032
Hi @lucydot, @yuxuanzhuang and @lorenzo-rovigatti did most of the heavy-lifting like proofreading and proposing suggestions to improve the codebase 👍 I also learned some useful tricks from the repo and discussions!

@braniii Great job putting together the codebase, as well as the documentation site, which, in addition to the paper, really helped me grasp the idea of msmhelper. I think I will do the same for my future work :)

Both installation steps and tutorials worked smoothly on my local computer, and I have no additional comments to the paper and codebase -- they all look good to me!
