+++ src/msmhelper/__init__.py
@@ -30,4 +30,4 @@ from .utils import (
     unique,
 )
 
-__version__ = '1.0.3'
+__version__ = '1.0.4'
