+++ CHANGELOG.md
@@ -16,6 +16,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [1.0.4] - 2023-05-22
+### Other changes:
+- Improvements of the README and the documentation, suggested by the JOSS reviewers @yuxuanzhuang and @lorenzo-rovigatti
+- Added zenodo setup
+- Added issue templates
+
+
 ## [1.0.3] - 2023-04-17
 ### Bugfix 🐛:
 - Fix that `mh.LumpedStateTraj.index_trajs` and `mh.LumpedStateTraj.index_trajs_flatten` return now index trajectories corresponding to the macrostate trajectories
@@ -165,7 +172,8 @@ Chapman-Kolmogorov test
 - Initial release
 
 
-[Unreleased]: https://github.com/moldyn/msmhelper/compare/v1.0.3...main
+[Unreleased]: https://github.com/moldyn/msmhelper/compare/v1.0.4...main
+[1.0.4]: https://github.com/moldyn/msmhelper/compare/v1.0.3...v1.0.4
 [1.0.3]: https://github.com/moldyn/msmhelper/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/moldyn/msmhelper/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/moldyn/msmhelper/compare/v1.0.0...v1.0.1
