### 2023-05-22 #5339 @braniii
Thank you all for your valuable feedback and reviews. @lucydot I've released a new version v1.0.4 to generate an zenodo archive of the final version: https://zenodo.org/record/7957966
