### 2023-05-24 #5339 @Kevin-Mattheus-Moerman
@lucydot thanks for editing this one, I'd like to point out some new editorialbot functionality:
`@editorialbot create post-review checklist`, it will create a helpful final checklist which you can use in the future prior to calling recomend-accept. I'll run it this time and will check some of them.
