### 2023-05-24 eb6fcd2
Add JOSS article to README.md
